removing handle_from_file from kernel32
