+++ pywincffi/kernel32/__init__.py
@@ -14,7 +14,7 @@ from pywincffi.kernel32.file import (
     UnlockFileEx)
 from pywincffi.kernel32.handle import (
     CloseHandle, GetStdHandle, GetHandleInformation, SetHandleInformation,
-    DuplicateHandle, handle_from_file)
+    DuplicateHandle)
 from pywincffi.kernel32.pipe import (
     CreatePipe, PeekNamedPipe, PeekNamedPipeResult, SetNamedPipeHandleState)
 from pywincffi.kernel32.process import (
