+++ pywincffi/kernel32/handle.py
@@ -17,29 +17,6 @@ from pywincffi.wintypes import HANDLE, wintype_to_cdata
 INVALID_HANDLE_VALUE = -1
 
 
-def handle_from_file(python_file):
-    """
-    Given a standard Python file object produce a Windows
-    handle object that be be used in Windows function calls.
-
-    :param file python_file:
-        The Python file object to convert to a Windows handle.
-
-    :return:
-        Returns a :class:`pywincffi.wintypes.HANDLE` for the provided
-        ``python_file``.
-    """
-    _, library = dist.load()
-    input_check("python_file", python_file, Enums.PYFILE)
-
-    # WARNING:
-    #   Be aware that passing in an invalid file descriptor
-    #   number can crash Python.  The input_check function
-    #   above should handle this for us by checking to
-    #   ensure the file descriptor is valid first.
-    return HANDLE(library.handle_from_fd(python_file.fileno()))
-
-
 def GetStdHandle(nStdHandle):
     """
     Retrieves a handle to the specified standard
