+++ tests/test_kernel32/test_handle.py
@@ -10,7 +10,7 @@ from pywincffi.core import dist
 from pywincffi.dev.testutil import TestCase
 from pywincffi.exceptions import InputError
 from pywincffi.kernel32 import (
-    GetStdHandle, CloseHandle, handle_from_file, GetHandleInformation,
+    GetStdHandle, CloseHandle, GetHandleInformation,
     SetHandleInformation, DuplicateHandle, GetCurrentProcess, CreateEvent)
 from pywincffi.wintypes import HANDLE
 
@@ -46,44 +46,7 @@ class TestGetStdHandle(TestCase):
         )
 
 
-class TestGetHandleFromFile(TestCase):
-    """
-    Tests for :func:`pywincffi.kernel32.handle_from_file`
-    """
-    def test_fails_if_not_a_file(self):
-        with self.assertRaises(InputError):
-            handle_from_file(0)
 
-    def test_fails_if_file_is_not_open(self):
-        fd, path = tempfile.mkstemp()
-        self.addCleanup(os.unlink, path)
-        test_file = os.fdopen(fd, "r")
-        test_file.close()
-
-        with self.assertRaises(InputError):
-            handle_from_file(test_file)
-
-    def test_opens_correct_file_handle(self):
-        fd, path = tempfile.mkstemp()
-        self.addCleanup(os.unlink, path)
-        os.close(fd)
-
-        test_file = open(path, "w")
-        handle = handle_from_file(test_file)
-
-        CloseHandle(handle)
-
-        # If CloseHandle() was passed the same handle
-        # that test_file is trying to write to the file
-        # and/or flushing it should fail.
-        try:
-            test_file.write("foo")
-            test_file.flush()
-        except (OSError, IOError, WindowsError) as error:
-            # EBADF == Bad file descriptor (because CloseHandle closed it)
-            self.assertEqual(error.errno, EBADF)
-        else:
-            self.fail("Expected os.close(%r) to fail" % fd)
 
 
 class TestGetHandleInformation(TestCase):
